fix: @roots/bud-terser
