--- sources/@roots/bud-terser/src/env.ts
@@ -1,11 +1,0 @@
-import type BudTerser from './extension.js'
-
-declare module '@roots/bud-framework' {
-  interface Bud {
-    terser: BudTerser
-  }
-
-  interface Modules {
-    '@roots/bud-terser': BudTerser
-  }
-}
