+++ sources/@roots/bud-terser/src/index.ts
@@ -10,7 +10,16 @@
  * @packageDocumentation
  */
 
-import './env.js'
-
 import BudTerser from './extension.js'
+
+declare module '@roots/bud-framework' {
+  interface Bud {
+    terser: BudTerser
+  }
+
+  interface Modules {
+    '@roots/bud-terser': BudTerser
+  }
+}
+
 export default BudTerser
